requirements.txt
aiogram==3.4.1
openai>=1.12.0
APScheduler==3.10.4
python-dotenv==1.0.1
SQLAlchemy==2.0.31
requests==2.32.3
pillow==10.4.0
fastapi==0.109.0
uvicorn==0.30.1
pydantic>=2.4.1,<2.6
aiohttp>=3.9.0,<3.10.0

